[Provider] Extract out ContentUriDefinition and put its code adding into class.
